+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.36.0
+
+##### ⚠ Breaking
+- Updated the native Android bridge to [Braze Android SDK 21.0.0](https://github.com/Appboy/appboy-android-sdk/blob/master/CHANGELOG.md#2100).
+- Updated the native iOS bridge to [Braze iOS SDK 4.4.4](https://github.com/Appboy/appboy-ios-sdk/blob/master/CHANGELOG.md#444).
+- Removed `setAvatarImageUrl()`.
+- Removed `logContentCardsDisplayed`. This method was not part of the recommended Content Cards integration and can be safely removed.
+
 # 1.35.1
 
 ##### Fixed
@@ -17,7 +25,7 @@
 # 1.34.0
 
 ##### ⚠ Breaking
-- Updated the native Android bridge to [Braze Android SDK 18.0.1](https://github.com/Appboy/appboy-android-sdk/blob/master/CHANGELOG.md#1801git).
+- Updated the native Android bridge to [Braze Android SDK 18.0.1](https://github.com/Appboy/appboy-android-sdk/blob/master/CHANGELOG.md#1801).
 
 ##### Fixed
 - Fixed an issue with Content Card types. Thanks @jtparret!